refactor(patching): one module per part of the patch engine
konfai/data/patching.py becomes the package konfai/data/patching: stage
(a chain stage as the engine sees it), budget (what a sweep may hold),
sweep (the streamed sweep), blend (patch weights), accumulate (the
reassembly), grid (the patch grids), manager (DatasetManager). Bodies
are moved verbatim; the manager reads the sweep tunables and the depth
rule through their modules, so a test's patch on the module reaches it.
