+++ konfai/data/patching/manager.py
@@ -14,33 +14,67 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Patch extraction, accumulation, and patch-combination helpers for KonfAI.
 
-:class:`DatasetManager` is one case's read side: it plans a chain against the stored volume
-(patch and region reads, the streamed replay through unmaterialized ``Save`` caches) and serves
-the DataLoader. Writing that chain to disk for the dataset-preparation workflow is the job of
-:class:`~konfai.data.materialize.CaseMaterializer`, which drives the plan/replay API defined here.
-"""
+"""The manager of one case: its chain, its grid, its reads and its writes."""
 
 import contextlib
 import copy
-import hashlib
-import itertools
-import queue
-import threading
 import warnings
-from abc import ABC, abstractmethod
-from collections.abc import Callable, Iterable, Iterator, Sequence
-from concurrent.futures import Future, ThreadPoolExecutor
-from dataclasses import dataclass, field
-from itertools import pairwise
-from typing import Any, NamedTuple, Protocol, TypeGuard, cast
+from collections.abc import Iterable, Iterator, Sequence
+from typing import Any, cast
 
 import numpy as np
 import torch
-import torch.nn.functional as F
 
-from konfai.data.augmentation import DataAugmentation, DataAugmentationsList
+from konfai.data.augmentation import DataAugmentationsList
+from konfai.data.patching import budget
+from konfai.data.patching import sweep as sweep_module
+from konfai.data.patching.budget import (
+    _PLATEAU_READ_MARGIN,
+    _STREAM_STAT_KEYS,
+    _STREAM_STATS,
+    _SWEEP_ELEMENT_BYTES,
+    _SWEEP_SLAB_ROWS_DEVICE,
+    _SWEEP_TILE_MARGIN,
+    _UNRESOLVED,
+)
+from konfai.data.patching.grid import DatasetPatch
+from konfai.data.patching.stage import (
+    _MAX_HALO_FRACTION,
+    AugmentedStage,
+    Stage,
+    _drawn_from,
+    _halo_radii,
+    _HaloPull,
+    _is_draw,
+    _ReadStagePlan,
+    _RemapPull,
+    _stage_name,
+)
+from konfai.data.patching.sweep import (
+    SWEEP_CLOCK,
+    BlockReads,
+    RegionWriter,
+    SweepSegment,
+    _channel_first_block,
+    _cubic_tile,
+    _HostLanding,
+    _open_sweep_stream,
+    _PatchStreamSource,
+    _PendingSweep,
+    _plateau_rows,
+    _pull_block_spans,
+    _ReadAhead,
+    _shares_h5_file,
+    _span_voxels,
+    _stage_failure,
+    _sweep_header,
+    _sweep_resident_regions,
+    _sweep_targets,
+    _SweepMember,
+    _WriteBehind,
+    save_destination,
+)
 from konfai.data.transform import (
     Expand,
     LocalityKind,
@@ -51,1927 +85,11 @@ from konfai.data.transform import (
     split_expand,
     stat_seed_valid,
 )
-from konfai.utils.budget import format_bytes, peak_resident_bytes, reset_resident_peak, resident_bytes
-from konfai.utils.clock import SweepClock
-from konfai.utils.config import apply_config, config
-from konfai.utils.dataset import Attribute, Dataset, DataStream, as_channel_first
+from konfai.utils.budget import format_bytes
+from konfai.utils.dataset import Attribute, Dataset
 from konfai.utils.dataset import chunk_hull_voxels as _chunk_hull_voxels
-from konfai.utils.errors import ConfigError, DatasetManagerError, PatchError, TransformError
-from konfai.utils.runtime import preserved_rng, rank_cpu_share, seed_all
-from konfai.utils.utils import (
-    OverlapSpec,
-    best_sweep_axis,
-    concretize_patch_size,
-    env_flag,
-    free_axis_rounding,
-    get_module,
-    get_patch_slices_from_shape,
-    resolve_overlap,
-)
-
-# How far a halo may reach, as a fraction of the patch it surrounds. See DatasetManager._affords_halo.
-_MAX_HALO_FRACTION = 0.5
-#: The whole-volume statistics every store can serve (``Dataset.read_data_statistics``), by the key a
-#: GLOBAL_STAT stage declares: what the plan checks against without reading a voxel, and what the
-#: seed reads. The per-channel figures are what a vector-valued quantity needs, where pooling the
-#: components into one number describes nothing.
-_STREAM_STATS = {
-    "Min": "min",
-    "Max": "max",
-    "Mean": "mean",
-    "Std": "std",
-    "MinPerChannel": "min_per_channel",
-    "MaxPerChannel": "max_per_channel",
-    "MeanPerChannel": "mean_per_channel",
-    "StdPerChannel": "std_per_channel",
-}
-_STREAM_STAT_KEYS = frozenset(_STREAM_STATS)
-
-# Rows per Save-sweep region: what bounds the materialization to a window while the composed region
-# reads stay chunk-friendly. A declared memory_budget can only LOWER the height (see
-# DatasetManager._sweep_rows); this constant is the cap and the no-budget default on a CPU. The
-# VOLUME it allows is what DatasetManager._sweep_tile then shapes into the block actually read.
-SWEEP_SLAB_ROWS = 64
-
-# "not looked up yet", where None is itself an answer (a store with no read granularity to state).
-_UNRESOLVED = object()
-
-# The bytes each element travels as through a sweep (float32). What a sweep holds in those elements
-# is _sweep_resident_regions, and DatasetManager.sweep_block_bytes prices it.
-#
-# What a streamed case holds beside its regions, whatever they are: the manager's state, the chain's
-# stage objects, the store handles, the allocator's slack. Measured as the resident set above the
-# interpreter floor that the priced regions do not account for, on the memory-limit cohort: 25 MiB
-# (a bare Write), 8 (Gradient), 11 (Resample), 27 (Resample then Gradient). A floor the sizing cannot
-# lower, so the TRANSFORM plan's header states it instead of leaving it to be found in a resident set.
-SWEEP_ENGINE_FLOOR_BYTES = 32 << 20
-#: The most blocks a sweep keeps in flight, whatever the budget leaves room for: past a second
-#: one the jitter it absorbs is already absorbed (DatasetManager._sweep_depth).
-_SWEEP_MAX_DEPTH = 3
-#: The region height cap when the chain runs on a GPU (bounded by free device memory as well).
-_SWEEP_SLAB_ROWS_DEVICE = 256
-#: How far above the floor a decomposition's reads may sit for its height to count as the plateau
-#: (:func:`_plateau_rows`). Measured on the prep's appearance fold, a native volume resampled
-#: through a field onto a 514x1331x1775 grid, reads against height as a multiple of the floor:
-#: 5 rows 1.79x, 10 1.40, 20 1.19, 41 1.09, 82 1.05, 169 1.02, 514 1.00. The curve is a knee, so
-#: anything from a few per cent to a tenth names the same height; five was measured to hold the
-#: fold at 17.4 GiB where an uncapped one held 28.7, for 56.9 -> 58.4 s of wall clock.
-_PLATEAU_READ_MARGIN = 0.05
-#: How much less a cubic block must read for the sweep to take it (``DatasetManager._sweep_tile``).
-#: A sheared map measures 0.61 on a 513x1331x1776 rigid+affine; an unsheared one, the margin alone.
-_SWEEP_TILE_MARGIN = 0.8
-_SWEEP_ELEMENT_BYTES = 4
-
-#: What a whole-volume fallback holds while a case is in flight (the assembled tensor plus one
-#: transform output), and the bytes each element travels as.
-#:
-#: Public because two callers must agree on the figure and neither owns it: the run-time budget check
-#: (``CaseMaterializer._enforce_fallback_budget``) refuses a case against it, and the TRANSFORM plan
-#: prints and enforces the same number before a byte is written. A plan estimating differently from
-#: the run it describes is worse than no plan.
-FALLBACK_INFLIGHT_FACTOR = 2
-CASE_ELEMENT_BYTES = 4
-
-
-def _halo_radii(halo: tuple[int, ...], n_axes: int) -> list[int]:
-    """The per-axis radius a declared halo means, in array order (one radius covers every axis)."""
-    if not halo:
-        return [0] * n_axes
-    return [halo[k] if k < len(halo) else halo[-1] for k in range(n_axes)]
-
-
-class Stage(Protocol):
-    """One step of what a case's copy is made of, as the patch-streaming dispatcher sees it.
-
-    A copy is its group's transforms followed by the augmentations drawn for it, and streaming asks the
-    same three things of every step: what its output depends on, which source region a target patch
-    needs, and to run on one tensor. A ``Transform`` answers them as itself. An augmentation is
-    parameterised per case and per copy, so it answers them bound to one (see :class:`AugmentedStage`).
-    """
-
-    def patch_locality(self, cache_attribute: Attribute) -> PatchLocality: ...
-
-    def stream_region_source(
-        self,
-        name: str,
-        target_slices: tuple[slice, ...],
-        source_spatial_shape: list[int],
-        cache_attribute: Attribute,
-    ) -> list[slice]: ...
-
-    def write_stream_cache_attribute(
-        self, cache_attribute: Attribute, source_spatial_shape: list[int], name: str = ""
-    ) -> None: ...
-
-    def stream_region(
-        self, name: str, tensor: torch.Tensor, context: RegionContext, cache_attribute: Attribute
-    ) -> torch.Tensor: ...
-
-    def plan_region_reads(self, name: str, contexts: Sequence[RegionContext]) -> None: ...
-
-    def __call__(self, name: str, tensor: torch.Tensor, cache_attribute: Attribute) -> torch.Tensor: ...
-
-
-def _spatial(shape: object) -> list[int]:
-    """A folded shape as plain Python ints, whatever the stage returned it as.
-
-    ``transform_shape`` is user-facing: a stage that computes its target grid with torch or numpy
-    hands back that library's scalars, and they travel unnoticed until something outside Python asks
-    for a real int: a zarr store refusing "Expected an iterable of integers", a header holding
-    ``tensor(128)``. Normalising at the fold makes ``shapes`` hold what it is typed as.
-    """
-    return [int(extent) for extent in cast("Sequence[Any]", shape)]
-
-
-def _is_draw(stage: object) -> TypeGuard[DataAugmentation]:
-    """Whether this chain entry is an augmentation: a stage the manager binds to a copy."""
-    return isinstance(stage, DataAugmentation)
-
-
-@contextlib.contextmanager
-def _drawn_from(*key: object) -> Iterator[None]:
-    """Seed the global RNGs (random, numpy, torch on every device) from ``key`` for the duration,
-    then restore them. Two chains of one case (an image and its mask) hold different draw objects
-    and derive the same copies from the same key (the Expand's seed, the case's name, which draw
-    this is). ``blake2b``, not ``hash()``: string hashing is salted per process.
-    """
-    digest = hashlib.blake2b("|".join(str(part) for part in key).encode(), digest_size=4).digest()
-    with preserved_rng():
-        seed_all(int.from_bytes(digest, "big"))
-        yield
-
-
-def _stage_name(stage: Stage) -> str:
-    """What to CALL a stage in a message: a draw's own class, never the adapter that binds it.
-
-    Every refusal and every regime note names the stage that caused it, and 'AugmentedStage' names
-    nothing a user wrote: the point of saying which stage refused is lost if the answer is the
-    wrapper's name for all of them.
-    """
-    if isinstance(stage, AugmentedStage):
-        return type(stage.augmentation).__name__
-    return type(stage).__name__
-
-
-@dataclass(frozen=True)
-class AugmentedStage:
-    """One augmentation, bound to the case and the copy whose draw it carries.
-
-    An augmentation is parameterised per (case, copy); binding both makes it answer the Stage
-    protocol like a plain transform.
-    """
-
-    augmentation: DataAugmentation
-    index: int
-    a: int
-
-    def patch_locality(self, cache_attribute: Attribute) -> PatchLocality:
-        return self.augmentation.patch_locality(self.index, self.a, cache_attribute)
-
-    def stream_region_source(
-        self,
-        name: str,
-        target_slices: tuple[slice, ...],
-        source_spatial_shape: list[int],
-        cache_attribute: Attribute,
-    ) -> list[slice]:
-        # A draw is bound to (case index, copy), not to the case's NAME: the name a region stage
-        # needs to find its own per-case map means nothing to an augmentation.
-        del name
-        return self.augmentation.stream_region_source(self.index, self.a, target_slices, source_spatial_shape)
-
-    def stream_region(
-        self, name: str, tensor: torch.Tensor, context: RegionContext, cache_attribute: Attribute
-    ) -> torch.Tensor:
-        del cache_attribute  # a draw reads no case metadata; the place is what it may need
-        return self.augmentation.stream_region(name, self.index, self.a, tensor, context)
-
-    def plan_region_reads(self, name: str, contexts: Sequence[RegionContext]) -> None:
-        """A draw reads no companion volume beside its region: nothing to declare."""
-
-    def write_stream_cache_attribute(
-        self, cache_attribute: Attribute, source_spatial_shape: list[int], name: str = ""
-    ) -> None:
-        """An augmentation draws a copy of the case rather than restating its geometry: nothing to record."""
-
-    def stream_shape(self, shape: list[int]) -> list[int]:
-        """The spatial shape this copy's draw produces from ``shape`` (its slot in the shape fold)."""
-        return self.augmentation.stream_shape(self.index, self.a, shape)
-
-    def __call__(self, name: str, tensor: torch.Tensor, cache_attribute: Attribute) -> torch.Tensor:
-        return self.augmentation.compute(name, self.index, self.a, tensor)
-
-
-# The pull maps are callable dataclasses, not closures, because a plan crosses a process boundary:
-# the launcher plans every case, `mp.spawn` then pickles the workflow object whole, and a local
-# function cannot be pickled. Everything a plan memoizes must survive that trip.
-
-
-@dataclass(frozen=True)
-class _HaloPull:
-    """A halo stage's pull map: the region enlarged by the radius, clamped to the volume."""
-
-    radii: list[int]
-    shape: list[int]
-
-    def __call__(self, target: tuple[slice, ...]) -> list[slice]:
-        return [
-            slice(max(0, t.start - radius), min(extent, t.stop + radius))
-            for t, radius, extent in zip(target, self.radii, self.shape, strict=False)
-        ]
-
-
-@dataclass(frozen=True)
-class _RemapPull:
-    """An index-remap stage's pull map, bound to the case and the state the stages before it left.
-
-    The case NAME is bound here because a stage instance is shared by every case of a manager
-    (``DatasetManager`` hands the same transforms list to each), while a map read from a stored
-    transform or a reference header is per case. A pull that could not say which case it was for
-    would build one case's window from another case's map, and a window that is short does not
-    raise, it returns the fill.
-    """
-
-    remap: Callable[[str, tuple[slice, ...], list[int], Attribute], list[slice]]
-    shape: list[int]
-    attribute: Attribute
-    name: str = ""
-
-    def __call__(self, target: tuple[slice, ...]) -> list[slice]:
-        return self.remap(self.name, target, list(self.shape), Attribute(self.attribute))
-
-
-@dataclass(frozen=True)
-class _ReadStagePlan:
-    """One chain stage as the composed streamed read runs it: its declared kind, the spatial shapes
-    on either side, and (for a region stage) the pull map from a region of its output to the region
-    of its input it is computed from, bound to the case state the stages before it left.
-
-    ``run_pull``, when set, is the pull the RUN walks instead: a stage that sizes its windows from
-    the data it reads (a declared field) measures there, while ``pull`` stays headers-only for the
-    plan's pricing: the estimator must never read a voxel."""
-
-    kind: LocalityKind
-    in_shape: tuple[int, ...]
-    out_shape: tuple[int, ...]
-    pull: Callable[[tuple[slice, ...]], list[slice]] | None
-    run_pull: Callable[[tuple[slice, ...]], list[slice]] | None = None
-
-    def region_context(self, source: Sequence[slice], target: Sequence[slice]) -> RegionContext:
-        """Where one region of this stage sits: the part of its input read, the part of its output due."""
-        return RegionContext(tuple(source), tuple(target), tuple(self.in_shape))
-
-
-def device_capped_budget(budget_bytes: float | None, device: "torch.device | None") -> float | None:
-    """The budget, capped at what ``device`` can actually hold.
-
-    The memory budget is declared in HOST bytes -- ``auto`` measures node RAM -- but on a GPU
-    chain the working sets it sizes (swept slabs, whole-volume fallbacks, a reduction's member
-    regions) live in VRAM. A 64G budget on a 16 GB card is then not a budget, it is a promise of
-    an OOM. Half of what the card can give THIS process: the halving is the slack that covers
-    allocations arriving after the reading, since a fold sized once can run for hours.
-
-    What the card can give this process is the free memory plus what this process's own allocator
-    is already sitting on, because a cached block is memory the next allocation reuses rather than
-    asks the driver for. Reading the free memory alone made the answer depend on WHEN it was read:
-    the same fold, on the same idle card, sized its regions at 47 rows under 4.99 GiB in one run
-    and 65 rows under 11.58 GiB in another, the difference being how much the process had already
-    reserved by the time the fit ran. Region height decides the whole read plan, so a sizing that
-    moves with the moment is a run whose cost cannot be reproduced or reasoned about.
-    """
-    if device is None or device.type != "cuda" or not torch.cuda.is_available():
-        return budget_bytes
-    free, _total = torch.cuda.mem_get_info(device)
-    vram = (free + torch.cuda.memory_reserved(device)) * 0.5
-    return vram if budget_bytes is None or budget_bytes <= 0 else min(budget_bytes, vram)
-
-
-@dataclass(frozen=True)
-class HeldMeter:
-    """What one scope of work HELD, read by the instrument the route it runs on has.
-
-    A GPU chain has the device allocator, which counts what is in use; a host chain has the kernel's
-    resident high-water mark, which counts what the allocator is sitting on as well -- and that is
-    the better figure of the two here, since the kernel kills on resident memory. Both answer the
-    same question, so a caller asks one thing and never branches on which it got.
-    """
-
-    _peak: Callable[[], int | None]
-    _baseline: int
-
-    def held(self) -> int | None:
-        """Bytes held above where the scope started, or ``None`` if the instrument went quiet."""
-        peak = self._peak()
-        return None if peak is None else max(0, peak - self._baseline)
-
-
-def open_held_meter(device: "torch.device | None") -> HeldMeter | None:
-    """Start measuring what the next scope holds, or ``None`` where nothing can measure it.
-
-    Reading is a high-water mark either way, so what comes back bounds the NEXT scope of the same
-    shape from above -- which is what makes a measured figure usable for sizing the ones that follow
-    (:meth:`Predictor._accumulate_device` reads a forward the same way).
-    """
-    if device is not None and device.type == "cuda" and torch.cuda.is_available():
-        try:
-            torch.cuda.synchronize(device)
-            torch.cuda.reset_peak_memory_stats(device)
-            baseline = int(torch.cuda.memory_allocated(device))
-        except Exception:  # nosec B110 - an unreadable instrument measures nothing, and says so
-            return None
-
-        def device_peak() -> int | None:
-            try:
-                torch.cuda.synchronize(device)
-                return int(torch.cuda.max_memory_allocated(device))
-            except Exception:  # nosec B110 - see above
-                return None
-
-        return HeldMeter(device_peak, baseline)
-    if not reset_resident_peak():
-        return None
-    resident = resident_bytes()
-    return None if resident is None else HeldMeter(peak_resident_bytes, int(resident))
-
-
-def save_destination(save: Save, default_dataset: Dataset, default_group: str) -> tuple[Dataset, str]:
-    """The dataset and group a :class:`Save` caches into, the manager's own when it names none.
-
-    Public because a planner has to resolve a destination exactly as the engine will: one that probes
-    a store the run does not open has verified nothing.
-    """
-    destination = save.destination
-    # No destination of its own: the Save caches into the manager's dataset, whose write format is
-    # not this stage's to redecorate: a pyramid asked here would silently not happen.
-    if destination is None and save.scale_factors:
-        raise ConfigError(
-            f"A '{type(save).__name__}' asks for a pyramid but names no dataset of its own.",
-            "scale_factors describes a store this stage writes, so give it one:"
-            " Write: {dataset: ./Out:omezarr, scale_factors: [4]}.",
-        )
-    return destination or default_dataset, save.group or default_group
-
-
-class RegionWriter:
-    """Region writes into streams opened at their first block: the one sweep loop of the three
-    engines (a Save's sweep, an Expand's shared pass, a Reduce).
-
-    ``write`` opens the stream for a key on its first block (through ``open_stream``, so a refusal
-    surfaces where the caller can say why) and writes the block; ``close`` publishes every stream,
-    ``abort`` drops every partial entry and is safe after a failure or an interrupt.
-
-    Writes here are synchronous. Whether a caller should overlap them with its compute depends on
-    where the compute runs, since the store's encoder competes for the host's cores: measured on a
-    300^3 float32 case to OME-Zarr, a host chain writing one slab behind was 50 % slower than
-    writing in place, and a host resample of a 513x1331x1776 case stayed at 13.0 s either way,
-    where the same resample on a GPU went from 8.2 s to 6.2 s with the write one block behind.
-    :class:`_WriteBehind` is what a sweep wraps this in; nothing here assumes it.
-    """
-
-    def __init__(self, open_stream: Callable[[Any, np.ndarray, Attribute], DataStream]) -> None:
-        self._open_stream = open_stream
-        self._streams: dict[Any, DataStream] = {}
-
-    @property
-    def opened(self) -> set[Any]:
-        return set(self._streams)
-
-    def write(self, key: Any, region: tuple[slice, ...], block: np.ndarray, header: Attribute) -> None:
-        """Write ``block`` at ``region`` (channel slice included) into ``key``'s stream, opened on
-        its first block with ``header``; a destination that refuses the stream raises there."""
-        if key not in self._streams:
-            stream = self._open_stream(key, block, header)
-            stream.__enter__()
-            self._streams[key] = stream
-        self._streams[key].write_slice(region, block)
-
-    def close(self) -> None:
-        """Publish every stream."""
-        for stream in self._streams.values():
-            stream.close()
-
-    def abort(self, error: BaseException | None = None) -> None:
-        """Drop every partial entry."""
-        for stream in self._streams.values():
-            stream.abort(error)
-        self._streams.clear()
-
-
-@dataclass(frozen=True)
-class _PatchStreamSource:
-    """What a copy's patches are read from, and what runs on them once read. ``stage_plans`` mirrors
-    ``stages`` one to one (the region plans compose, each pulling through the one before it);
-    ``pending_sweeps`` are the unsatisfied :class:`Save` caches to materialize before any patch
-    flows; ``entry`` is the case's own name, or the copy's behind an :class:`Expand`.
-    """
-
-    dataset: Dataset
-    group: str
-    entry: str
-    shape: list[int]
-    stages: list[Stage]
-    stage_plans: tuple[_ReadStagePlan, ...]
-    pending_sweeps: tuple["_PendingSweep", ...] = ()
-
-    @property
-    def region_index(self) -> int | None:
-        """The first region stage, or ``None`` for an exact-patch chain."""
-        for index, plan in enumerate(self.stage_plans):
-            if plan.kind.is_region:
-                return index
-        return None
-
-
-@dataclass(frozen=True)
-class _PendingSweep:
-    """One unsatisfied :class:`Save` and the segment that feeds it: the sweep reads each block of the
-    Save's space through the segment (re-planned at sweep time) and region-writes it to
-    ``destination``. ``entry``/``source_entry`` are the destination and source entry names (the
-    copy's own behind an :class:`Expand`); ``copy_stage_start`` is where the per-copy part of the
-    segment begins (``len(stages)`` when it is entirely shared); ``stage_plans`` are the probe-time
-    plans, kept for the shared-pass classification only."""
-
-    destination: Dataset
-    group: str
-    entry: str
-    stages: list[Stage]
-    source_dataset: Dataset
-    source_group: str
-    source_entry: str
-    source_shape: list[int]
-    out_spatial: tuple[int, ...]
-    # The case state the segment replays from (copied per block).
-    base_attributes: Attribute = field(repr=False)
-    copy_stage_start: int = 0
-    stage_plans: tuple[_ReadStagePlan, ...] = ()
-
-    @property
-    def tail_stages(self) -> tuple[Stage, ...]:
-        """The per-copy part of the segment: what a copy applies to a block the copies share."""
-        return tuple(self.stages[self.copy_stage_start :])
-
-    @property
-    def tail_plans(self) -> tuple[_ReadStagePlan, ...]:
-        """The probe-time plans of :attr:`tail_stages`."""
-        return self.stage_plans[self.copy_stage_start :]
-
-
-@dataclass(frozen=True)
-class _SweepMember:
-    """One stream a sweep writes: the key it is reported under, the :class:`Save` it lands in, the
-    case state its plan leaves (its header), and the per-copy tail it applies to each block.
-
-    ``stages``/``stage_plans`` name the tail in the same shape as a :class:`_PatchStreamSource`'s
-    chain, so the block dispatch is the one the stages before the marker take. Every tail plan is
-    POINTWISE (``CaseMaterializer._pointwise_tail``): the block a member is handed IS its region,
-    which is what lets the copies share one read pass."""
-
-    key: Any
-    sweep: _PendingSweep
-    evolved: Attribute = field(repr=False)
-    stages: tuple[Stage, ...] = ()
-    stage_plans: tuple[_ReadStagePlan, ...] = ()
-
-    def region_spans(self, target: tuple[slice, ...]) -> list[list[slice]]:
-        """The region each tail stage reads for a block of the landing: the block itself, for all of
-        them, a pointwise stage pulling exactly what it is handed."""
-        return [list(target) for _ in range(len(self.stages) + 1)]
-
-
-def _sweep_targets(spatial: list[int], tile: Sequence[int]) -> Iterator[tuple[slice, ...]]:
-    """The sweep's regions: ``tile``-sized blocks of the landing, innermost axis fastest.
-
-    The order is the source's: consecutive blocks differ on the axis stored contiguously, so the
-    chunks one block decodes are the chunks the next one reads.
-    """
-    steps = [max(1, int(step)) for step in tile]
-    for corner in itertools.product(*(range(0, extent, step) for extent, step in zip(spatial, steps, strict=True))):
-        yield tuple(
-            slice(start, min(start + step, extent)) for start, step, extent in zip(corner, steps, spatial, strict=True)
-        )
-
-
-def _cubic_tile(spatial: list[int], voxels: int, align: int) -> list[int]:
-    """The block of at most ``voxels`` closest to a cube inside ``spatial``, aligned to ``align``.
-
-    ``align`` keeps a block a whole number of store chunks wide, so a region write never becomes a
-    read-modify-write (:func:`konfai.utils.dataset._store_chunks`); an axis shorter than one step is
-    taken whole. Why a cube: :meth:`DatasetManager._sweep_tile`.
-    """
-    tile = [max(1, int(extent)) for extent in spatial]
-    budget = float(max(1, voxels))
-    # Shortest axis first: an axis under the ideal side takes its whole extent and hands its slack
-    # to the others, which only raises the side, so once one axis is over it every later one is too.
-    free = sorted(range(len(spatial)), key=lambda axis: tile[axis])
-    for taken, axis in enumerate(free):
-        side = budget ** (1.0 / (len(free) - taken))
-        if tile[axis] > side:
-            for remaining in free[taken:]:
-                tile[remaining] = max(1, min(tile[remaining], int(side)))
-            break
-        budget /= tile[axis]
-    return [
-        extent if extent >= spatial[axis] or extent < align else extent - extent % align
-        for axis, extent in enumerate(tile)
-    ]
-
-
-def _pull_block_spans(
-    spatial: list[int], tile: Sequence[int], plans: Sequence["_ReadStagePlan"]
-) -> Iterator[list[slice]]:
-    """The source window each block of a decomposition into ``tile`` reads, from the plans' own pull
-    maps: closed form, no voxel read."""
-    for target in _sweep_targets(spatial, tile):
-        span = list(target)
-        for plan in reversed(plans):
-            span = list(plan.pull(tuple(span))) if plan.pull is not None else span
-        yield span
-
-
-def _span_voxels(span: Sequence[slice]) -> int:
-    return int(np.prod([max(0, part.stop - part.start) for part in span], dtype=np.int64))
-
-
-class BlockReads(NamedTuple):
-    """What one decomposition's blocks read, in the currencies the sizing spends.
-
-    Every consumer of the enumeration wants an aggregate of it and none wants the blocks, so it is
-    walked once and the aggregates are kept: the widest window one block materialises, the widest
-    hull the store decodes to serve one, and what the whole decomposition reads. Walking it per
-    consumer made the plan spend 96% of its time in ``Resample.stream_region_source``.
-    """
-
-    widest_pull: int
-    widest_hull: int
-    total: int
-
-    @property
-    def widest_excess(self) -> int:
-        """What the widest read materialises ABOVE the window it asked for: zero on a store that
-        serves exactly what it is asked for."""
-        return max(0, self.widest_hull - self.widest_pull)
-
-
-def _pull_block_voxels(spatial: list[int], tile: Sequence[int], plans: Sequence["_ReadStagePlan"]) -> Iterator[int]:
-    """The source voxels each block of a decomposition into ``tile`` materialises. Their sum is what
-    the decomposition reads, their largest what one block of it holds."""
-    return (_span_voxels(span) for span in _pull_block_spans(spatial, tile, plans))
-
-
-def _plateau_rows(
-    spatial: list[int], plans: Sequence["_ReadStagePlan"], tolerance: float = _PLATEAU_READ_MARGIN
-) -> int | None:
-    """The shortest region height whose decomposition already reads what the tallest one reads,
-    within ``tolerance``: the height past which taller regions pull no fewer source voxels.
-
-    Closed form, from the chain's own pull maps (:func:`_pull_block_voxels`): no voxel is read, so
-    it is answerable at plan time for any chain, whatever the YAML declares. ``None`` when nothing
-    pulls (no plans), where every height reads the same and the question does not arise.
-
-    This is a CAP, never a target: below it a region re-reads what its neighbour already pulled,
-    above it the reads are the same and only the working set grows. What a budget then affords is
-    the caller's to search for, downward.
-    """
-    if not plans:
-        return None
-    total = max(1, int(spatial[0]))
-    floor = sum(_pull_block_voxels(spatial, [total, *spatial[1:]], plans))
-    if floor <= 0:
-        return None
-    allowed = floor * (1.0 + max(0.0, float(tolerance)))
-    # Geometric, so the scan costs O(log) evaluations however tall the volume is.
-    height, ladder = max(1, total // 256), []
-    while height < total:
-        ladder.append(height)
-        height = max(height + 1, int(height * 1.5))
-    ladder.append(total)
-    for candidate in ladder:
-        if sum(_pull_block_voxels(spatial, [candidate, *spatial[1:]], plans)) <= allowed:
-            return candidate
-    return total
-
-
-def _sweep_pipeline_depth() -> int:
-    """How many blocks a sweep keeps in flight beside the one it is transforming: one ahead and one
-    behind, or none when the rank owns a single core (``OMP_NUM_THREADS=1`` means a serial run)."""
-    return 1 if rank_cpu_share() > 1 else 0
-
-
-def _sweep_resident_regions(depth: int) -> tuple[int, int]:
-    """How many pulled source regions and how many landed blocks a sweep of pipeline ``depth`` holds
-    at once: the region the chain is running on, and, pipelined, the ``depth`` queued ahead of it
-    plus the one the reader holds while the queue is full; against the block being landed and,
-    pipelined, the one being written behind it."""
-    return (1, 1) if depth < 1 else (depth + 2, 2)
-
-
-class SweepSegment(NamedTuple):
-    """One segment the streamed route sweeps: where it reads from, what it lands, how it pulls."""
-
-    dataset: Dataset
-    group: str
-    entry: str
-    source_shape: list[int]
-    landing: list[int]
-    plans: tuple["_ReadStagePlan", ...]
-
-    @property
-    def channels(self) -> int:
-        return int(self.source_shape[0])
-
-
-#: This rank's sweeps, summed over the cases it ran: the unit the report is about.
-SWEEP_CLOCK = SweepClock()
-
-
-class _ReadAhead:
-    """One block's read running while the previous one is transformed and written.
-
-    The consumer drains IN ORDER, so a stage object is touched by one thread at a time: the read
-    stages by the producer, the tail by the consumer, disjoint by construction. At most ``depth``
-    blocks wait, and a consumer that leaves early stops the producer.
-    """
-
-    _DONE = object()
-
-    def __init__(self, blocks: Iterator[Any], depth: int) -> None:
-        self._blocks = blocks
-        self._depth = depth
-        self._queue: queue.Queue = queue.Queue(maxsize=max(1, depth))
-        self._stop = threading.Event()
-        self._thread = threading.Thread(target=self._produce, name="konfai-sweep-read", daemon=True)
-
-    def __enter__(self) -> Iterator[Any]:
-        if self._depth < 1:
-            return self._blocks
-        self._thread.start()
-        return self._consume()
-
-    def __exit__(self, exc_type, value, traceback) -> None:
-        if self._depth < 1:
-            return
-        # To the end marker, whether the consumer stopped early or an error cut it short: a producer
-        # blocked on a full queue cannot reach its own end, and cannot then be joined.
-        self._stop.set()
-        while self._queue.get()[0] is not _ReadAhead._DONE:
-            pass
-        self._thread.join()
-
-    def _produce(self) -> None:
-        try:
-            for block in self._blocks:
-                self._queue.put((block, None))
-                if self._stop.is_set():
-                    break
-        except BaseException as error:  # re-raised in the consumer's thread, where the sweep is
-            self._queue.put((None, error))
-        self._queue.put((_ReadAhead._DONE, None))
-
-    def _consume(self) -> Iterator[Any]:
-        while True:
-            block, error = self._queue.get()
-            if block is _ReadAhead._DONE:
-                self._queue.put((block, None))  # left for __exit__, which drains to it
-                return
-            if error is not None:
-                raise error
-            yield block
-
-
-class _HostLanding:
-    """The host buffers a device chain's blocks come home into, reused rather than reallocated.
-
-    A fresh allocation per block pays first-touch on pages the transfer overwrites whole, so the
-    pageable copy faults page by page: measured on a 216 MiB block, 2.4 GiB/s into a new buffer
-    against 10.1 GiB/s into one already resident. Two slots, because the writer holds exactly one
-    block while the sweep fills the next, and that is what the sweep already keeps live.
-
-    A block already on the host is handed over as it is: copying one would cost the host chain
-    exactly what this saves on the device one.
-    """
-
-    _SLOTS = 2
-
-    def __init__(self) -> None:
-        self._slots: list[torch.Tensor] = [torch.empty(0, dtype=torch.uint8) for _ in range(_HostLanding._SLOTS)]
-        self._next = 0
-
-    def take(self, tensor: torch.Tensor) -> np.ndarray:
-        """``tensor``'s values on the host, in a buffer this owns (or its own, if it is host-side)."""
-        if tensor.device.type == "cpu":
-            return tensor.numpy()
-        slot, self._next = self._next, (self._next + 1) % _HostLanding._SLOTS
-        needed = tensor.nelement() * tensor.element_size()
-        if self._slots[slot].nelement() < needed:
-            self._slots[slot] = torch.empty(needed, dtype=torch.uint8)
-        landing = self._slots[slot][:needed].view(tensor.dtype).view(tensor.shape)
-        landing.copy_(tensor)
-        return landing.numpy()
-
-
-class _WriteBehind:
-    """A sweep's :class:`RegionWriter`, driven from one thread that is not the sweep's.
-
-    One worker and one outstanding write, so the order stays the sweep's and the store's own
-    encoder keeps its concurrency. EVERY call goes to that thread, not only the writes: the h5
-    backend holds a per-file ``RLock`` across a stream's whole life, and a release from a thread
-    that did not take it raises.
-    """
-
-    def __init__(self, writer: RegionWriter, depth: int) -> None:
-        self._writer = writer
-        self._pool = ThreadPoolExecutor(max_workers=1, thread_name_prefix="konfai-sweep-write") if depth else None
-        self._pending: Future | None = None
-
-    def write(self, key: Any, region: tuple[slice, ...], block: np.ndarray, header: Attribute) -> None:
-        if self._pool is None:
-            self._timed_write(key, region, block, header)
-            return
-        self.flush()  # one outstanding write, so the order is the sweep's and the memory is bounded
-        self._pending = self._pool.submit(self._timed_write, key, region, block, header)
-
-    def _timed_write(self, key: Any, region: tuple[slice, ...], block: np.ndarray, header: Attribute) -> None:
-        with SWEEP_CLOCK.phase("write"):
-            self._writer.write(key, region, block, header)
-
-    def flush(self) -> None:
-        """Wait for the outstanding write, raising what it raised."""
-        pending, self._pending = self._pending, None
-        if pending is not None:
-            pending.result()
-
-    def close(self) -> set[Any]:
-        """Publish every stream and answer which keys were written."""
-        self.flush()
-        return self._on_writer(self._publish)
-
-    def abort(self, error: BaseException) -> None:
-        """Drop every partial entry, whatever the outstanding write did."""
-        with contextlib.suppress(Exception):
-            self.flush()
-        self._on_writer(self._writer.abort, error)
-
-    def shutdown(self) -> None:
-        if self._pool is not None:
-            self._pool.shutdown(wait=True)
-
-    def _publish(self) -> set[Any]:
-        written = self._writer.opened
-        with SWEEP_CLOCK.phase("write"):
-            self._writer.close()
-        return written
-
-    def _on_writer(self, work: Callable[..., Any], *args: Any) -> Any:
-        return work(*args) if self._pool is None else self._pool.submit(work, *args).result()
-
-
-def _torch_dtype_hint(error: BaseException) -> str | None:
-    """The config change that answers ``error``, when it is a dtype torch has no kernel for.
-
-    torch ships none for several dtypes a store legitimately holds: ``uint16`` is what microscopy
-    writes, and torch implements for it neither comparison, nor flip, nor arithmetic, nor scalar
-    fill. A chain that touches such a payload raises deep inside a stage with nothing but the
-    missing operator's name, where what the reader needs is the dtype and the line that fixes it.
-    """
-    text = str(error)
-    if not isinstance(error, NotImplementedError) or "not implemented for" not in text or text.count("'") < 2:
-        return None
-    return (
-        f"torch has no kernel for a '{text.rsplit(chr(39), 2)[-2]}' payload: put a cast at the head"
-        " of the chain (TensorCast, dtype float32, or int32 to keep every value exact) so the stages"
-        " work in a dtype torch implements. The store keeps its own."
-    )
-
-
-@contextlib.contextmanager
-def _stage_failures_explained() -> Iterator[None]:
-    """Re-raise what a chain raised with the config change that answers it, where there is one."""
-    try:
-        yield
-    except NotImplementedError as error:
-        hint = _torch_dtype_hint(error)
-        if hint is None:
-            raise
-        raise TransformError(str(error), hint) from error
-
-
-def _stage_failure(error: BaseException) -> str:
-    """What a chain raised, with the hint appended where one applies."""
-    hint = _torch_dtype_hint(error)
-    return f"{type(error).__name__}: {error}" + (f". {hint}" if hint else "")
-
-
-def _shares_h5_file(source: Dataset, destination: Dataset) -> bool:
-    """Whether a read of ``source`` and a stream into ``destination`` can open the same h5 file.
-    By root: a directory root holds one file per case, and the same root is the same file."""
-    return (
-        source.file_format == "h5" and destination.file_format == "h5" and source.store_root == destination.store_root
-    )
-
-
-def _sweep_header(evolved: Attribute, scope: Attribute, keys_before: set[str]) -> Attribute:
-    """The header a sweep publishes: the plan-time case state, completed by what the chain's
-    stages added in ``__call__``: the same keys the whole-volume pass would leave at the Save."""
-    attributes = Attribute(evolved)
-    for key in scope.keys():
-        if key not in keys_before and key not in attributes:
-            attributes[key] = scope[key]
-    return attributes
-
-
-def _channel_first_block(block: np.ndarray, spatial: list[int], header: Attribute, what: str) -> np.ndarray:
-    """The block a region write ships: channel-first, single-channel where the header says so
-    (:func:`as_channel_first`, the rule the whole-volume write applies), else refused.
-
-    Writing another rank anyway is the worst outcome available: the header would take the block's
-    first spatial extent for a channel count and publish a store of that many "channels", raising
-    nothing, while the whole-volume path returns the right rank: the two would silently disagree."""
-    block = as_channel_first(block, header)
-    if block.ndim != len(spatial) + 1:
-        raise PatchError(
-            f"{what} returned a rank-{block.ndim} block where the channel-first layout needs rank"
-            f" {len(spatial) + 1} (C, {', '.join(str(extent) for extent in spatial)}).",
-            "A transform that reduces the leading axis must keep it (`keepdim=True`), so a block"
-            " stays C[Z]YX; only then does a region write mean the same thing as the"
-            " whole-volume pass.",
-        )
-    return block
-
-
-def _open_sweep_stream(
-    sweep: _PendingSweep, block: np.ndarray, spatial: list[int], tile: Sequence[int], attributes: Attribute
-) -> DataStream:
-    """One region-write stream shaped for the sweep: the store chunks divide the block the sweep
-    writes (channels included), so no region write ever pays a read-modify-write."""
-    stream = sweep.destination.open_data_stream(
-        sweep.group,
-        sweep.entry,
-        [int(block.shape[0]), *spatial],
-        block.dtype,
-        attributes,
-        region_shape=[int(block.shape[0]), *(int(extent) for extent in tile)],
-    )
-    if stream is None:
-        raise PatchError(
-            f"destination '{sweep.destination.filename}' refused the region write of"
-            f" '{sweep.group}/{sweep.entry}' after accepting its plan.",
-            "h5 and omezarr always serve region writes; mha only with image geometry.",
-        )
-    return stream
-
-
-@dataclass(frozen=True)
-class PatchReadPlan:
-    """Precomputed slicing and padding instructions for one patch request."""
-
-    data_slices: tuple[slice, ...]
-    reflect_padding: tuple[int, ...]
-    constant_padding: tuple[int, ...]
-    concatenate_extend_slice: bool
-
-
-class PathCombine(ABC):
-    """Base class for overlap-aware weighting schemes applied during patch assembly."""
-
-    def __init__(self) -> None:
-        self.windows_1d: list[torch.Tensor]
-        self.overlap: int
-        self.overlaps: list[int]
-        self._data_per_device: dict[tuple[torch.device, torch.dtype], torch.Tensor] = {}
-
-    def set_patch_config(self, patch_size: list[int], overlap: int | list[int]) -> None:
-        self._data_per_device.clear()
-        overlaps = [overlap] * len(patch_size) if isinstance(overlap, int) else list(overlap)
-        self.overlap = max(overlaps)
-        self.overlaps = overlaps
-        # The per-patch weight is the outer product of one 1-D window per axis. It is separable by
-        # construction, so a per-axis partition of unity stays a partition of unity in N-D and overlapping
-        # patches blend without darkening: no distance map or explicit renormalisation loop is needed, and
-        # the trailing normalisation in Accumulator.assemble stays exact at the volume borders.
-        # Each axis tapers over its own overlap, so anisotropic overlaps blend correctly per axis.
-        #
-        # The factors are kept, not just their product: Accumulator normalises by a per-axis sum rather
-        # than a volume-sized buffer (see _weight_factors). Stored rather than re-derived, because the
-        # all-flat case below never calls _window_1d: re-deriving would turn its uniform count into a
-        # tapered weighting.
-        if all(o <= 0 for o in overlaps):
-            self.windows_1d = [torch.ones(size) for size in patch_size]
-            return
-        self.windows_1d = [
-            self._window_1d(size, axis_overlap) if axis_overlap > 0 else torch.ones(size)
-            for size, axis_overlap in zip(patch_size, overlaps, strict=True)
-        ]
-
-    @property
-    def data(self) -> torch.Tensor:
-        """The per-voxel window: the outer product of the per-axis factors.
-
-        Derived, not stored: it is a patch of floats that ``Network.init`` builds before the spawn and
-        every rank then unpickles (a 128^3 ModelPatch at overlap 16 pickles to 8 391 775 bytes with it,
-        1355 without), while assembly goes through the factors themselves (see ``_weight_factors``).
-        :meth:`weight` caches what it builds per device and dtype.
-        """
-        data = self.windows_1d[0]
-        for window in self.windows_1d[1:]:
-            data = data.unsqueeze(-1) * window
-        return data
-
-    @property
-    def selects(self) -> bool:
-        """Whether the window picks ONE patch per voxel (values 0 or 1) rather than weighting several.
-
-        A selection makes the kept regions a partition of the volume, so assembly writes each one
-        instead of accumulating a weighted sum: see Accumulator._blend.
-        """
-        return False
-
-    def window(self, dim: int, position: int, count: int) -> torch.Tensor:
-        """The 1-D window along ``dim`` for the patch at grid ``position`` of ``count`` on that axis.
-
-        A weighting is the same wherever the patch sits; a selection opens its border patches so the
-        kept bands still reach the volume edge (see :class:`Trim`).
-        """
-        del position, count
-        return self.windows_1d[dim]
-
-    def weight(self, tensor: torch.Tensor) -> torch.Tensor:
-        """The raw per-voxel window on ``tensor``'s device and dtype (cached per pair).
-
-        The window UNNORMALISED, as declared. Assembly does not go through it: Accumulator divides each
-        window by the total over the patches covering a voxel and applies that share instead, so the
-        quantity it multiplies by is always in [0, 1]. This stays for a caller that wants the window
-        itself.
-        """
-        key = (tensor.device, tensor.dtype)
-        if key not in self._data_per_device:
-            # Match the tensor dtype: the weight has no reason to carry more precision than the data it
-            # scales, and a float64 weight would upcast the whole (channels x volume) blend.
-            weight = self.data.to(device=tensor.device, dtype=tensor.dtype)
-            if weight.is_floating_point():
-                # A Gaussian tail underflows the target dtype (a 96^3 patch corner is ~6e-11: zero in
-                # fp16). Floor it at the smallest normal so a caller scaling by this window on its own
-                # does not silently drop those voxels.
-                weight = weight.clamp(min=torch.finfo(weight.dtype).tiny)
-            self._data_per_device[key] = weight
-        return self._data_per_device[key]
-
-    def __call__(self, tensor: torch.Tensor) -> torch.Tensor:
-        return self.weight(tensor) * tensor
-
-    @abstractmethod
-    def _window_1d(self, size: int, overlap: int) -> torch.Tensor:
-        """Return the 1-D blend weight along one axis (length ``size``, ``overlap`` voxels tapered per side)."""
-
-
-def blend_axes(patch_size: list[int]) -> list[int]:
-    """The blend window's axes for a ``patch_size``: every axis kept, a free (0) or singleton (1) axis
-    as a single broadcast entry.
-
-    ``Accumulator`` reads one window per spatial axis of the volume, and the window has to broadcast
-    against the patch at any axis position. Dropping the untiled axes shortens the list, so the
-    window is looked up on the wrong axis (or not at all).
-    """
-    return [size if size > 1 else 1 for size in patch_size]
-
-
-def blend_overlap(overlap: "int | float | str | list[int | float | str]", patch_size: list[int]) -> list[int]:
-    """Per-axis blend overlap for a concrete ``patch_size`` (int broadcast, ``%``/fraction resolved).
-
-    The blend has no volume extent: every axis whose patch is longer than one voxel is treated as tiled
-    (the untiled-axis-0 rule is already applied by the slicing plan), so an int is the same voxel
-    overlap on every such axis.
-    """
-    if isinstance(overlap, int):
-        return [overlap if size > 1 else 0 for size in patch_size]
-    return resolve_overlap(overlap, patch_size, [size + 1 for size in patch_size])
-
-
-class Mean(PathCombine):
-    """Uniform weighting: overlapping patches are plain-averaged by the assembly normalisation."""
-
-    def _window_1d(self, size: int, overlap: int) -> torch.Tensor:
-        return torch.ones(size)
-
-
-class Cosinus(PathCombine):
-    """Raised-cosine (sin**2) taper: a smooth partition of unity across the patch overlap."""
-
-    def _window_1d(self, size: int, overlap: int) -> torch.Tensor:
-        window = torch.ones(size)
-        # sin**2 ramp over the overlap; the neighbouring patch's cos**2 ramp is its complement, so the two
-        # sum to exactly one across the overlap. The +0.5 phase keeps the very edge > 0 so a single-patch
-        # border, whose share is then the whole of the weight there, recovers the raw value.
-        ramp = torch.sin((torch.arange(overlap, dtype=torch.float32) + 0.5) / overlap * (torch.pi / 2)) ** 2
-        window[:overlap] = ramp
-        window[size - overlap :] = ramp.flip(0)
-        return window
-
-
-class Trim(PathCombine):
-    """Selection instead of weighting: each voxel comes from the patch that holds it most centrally.
-
-    An interior patch keeps its central ``patch - overlap`` band, so the kept bands tile the axis
-    exactly and the weights already sum to one; the first and last patch of an axis open to the volume
-    edge. Every kept voxel therefore carries at least half the overlap of context on each side, where
-    the unweighted default keeps whichever patch wrote last: possibly one holding that voxel on its
-    own border with no context behind it, which is what a seam is.
-
-    The values are 0 or 1, so the patch is selected rather than averaged: a discrete output (a label
-    map, an argmax) survives reassembly, where any weighting would invent values between its classes.
-    """
-
-    @property
-    def selects(self) -> bool:
-        return True
-
-    def _window_1d(self, size: int, overlap: int) -> torch.Tensor:
-        if overlap >= size:
-            # Nothing central left to keep: trimming both sides would leave an empty band, and a patch
-            # that keeps nothing has no box to write. Keep the patch whole instead: the axis stops
-            # being a partition and falls back to last-write-wins on it, which is what it was before.
-            return torch.ones(size)
-        window = torch.zeros(size)
-        # Split an odd overlap so consecutive kept bands abut: k*stride + hi == (k+1)*stride + lo.
-        window[overlap // 2 : size - (overlap - overlap // 2)] = 1.0
-        return window
-
-    def window(self, dim: int, position: int, count: int) -> torch.Tensor:
-        overlap = self.overlaps[dim]
-        window = self.windows_1d[dim]
-        if overlap <= 0 or (position > 0 and position < count - 1):
-            return window
-        window = window.clone()
-        if position == 0:
-            window[: overlap // 2] = 1.0
-        if position == count - 1:
-            window[window.shape[0] - (overlap - overlap // 2) :] = 1.0
-        return window
-
-
-class Gaussian(PathCombine):
-    """nnU-Net-style Gaussian importance weighting.
-
-    Favours patch centres (where the model sees the most surrounding context), and down-weights the
-    borders, which suppresses seam artefacts. Not a partition of unity, but ``Accumulator.assemble``
-    normalises by the accumulated weight, so overlapping patches still form a correct weighted average.
-    """
-
-    def __init__(self, sigma_scale: float = 0.125) -> None:
-        super().__init__()
-        self.sigma_scale = sigma_scale
-
-    def _window_1d(self, size: int, overlap: int) -> torch.Tensor:
-        sigma = max(size * self.sigma_scale, 1e-6)
-        center = (size - 1) / 2
-        coords = torch.arange(size, dtype=torch.float32)
-        return torch.exp(-((coords - center) ** 2) / (2.0 * sigma**2))
-
-
-class Accumulator:
-    """Accumulate patch predictions and reassemble them into a full tensor."""
-
-    def __init__(
-        self,
-        patch_slices: list[tuple[slice, ...]],
-        patch_size: list[int],
-        patch_combine: PathCombine | None = None,
-        batch: bool = True,
-        sweep_axis: int = 0,
-    ) -> None:
-        # Which spatial axis the window slides along. The patch grid is emitted with this axis
-        # outermost, so a patch's arrival finalizes everything behind it on that axis and nothing
-        # else. 0 is what get_patch_slices_from_shape produces today.
-        self.sweep_axis = sweep_axis
-        self.patch_slices: list[tuple[slice, ...]] = []
-        self.shape = max([[v.stop for v in patch] for patch in patch_slices])
-
-        if patch_size is not None and not all(p == 0 for p in patch_size):
-            # The last patch of an axis is padded up to the patch size for the model, then cropped; a
-            # free axis (0) spans the full extent, so concretise it here or ``s.start + 0`` would
-            # collapse the axis to a zero-width slice.
-            concrete = [size if size > 0 else self.shape[dim] for dim, size in enumerate(patch_size)]
-            for patch in patch_slices:
-                slices = [slice(s.start, s.start + concrete[dim]) for dim, s in enumerate(patch)]
-                self.patch_slices.append(tuple(slices))
-        else:
-            self.patch_slices = patch_slices
-        self.patch_size = patch_size
-        self.patch_combine = patch_combine
-        self.batch = batch
-        self._n = 2 if batch else 1
-        self._count = len(patch_slices)
-        self._filled = 0
-        self._done = [False] * self._count
-        # Patches are blended into this running buffer as they arrive (see add_layer), instead of
-        # being kept until assembly: holding every patch of a large multi-class case (e.g. ~70 patches
-        # of a 122-channel whole-body segmentation ≈ tens of GB) was the dominant reassembly RAM cost.
-        self._result: torch.Tensor | None = None
-        self._weighted: torch.Tensor | None = None
-        # Blend-weight geometry: fixed for the accumulator's life, so it survives _reset. The grid,
-        # the shares and the kept spans are all per (axis, start): sum(n_d) entries for a grid of
-        # prod(n_d) patches, and one dict lookup per patch and axis to find them.
-        self._geometry: tuple[list[list[torch.Tensor]], list[torch.Tensor]] | None = None
-        self._grid: list[dict[int, int]] | None = None
-        self._shares: dict[tuple[int, int, int, torch.dtype, torch.device], torch.Tensor | None] = {}
-        self._kept: dict[tuple[int, int], slice] = {}
-
-    def add_layer(self, index: int, layer: torch.Tensor) -> list[tuple[slice, torch.Tensor]]:
-        """Blend one patch in; returns the slabs this completes (none for the whole-volume base)."""
-        # Blend each patch straight into the running accumulator and drop the patch, rather than
-        # storing all patches for a single assemble() at the end. The overlap blend is a weighted sum,
-        # so accumulating incrementally is equivalent; re-adding an index is a no-op (last-write wins is
-        # not possible once blended, and the prediction pipeline adds each patch exactly once).
-        if self._done[index]:
-            return []
-        if self._result is None:
-            # Allocate to the ACTUAL volume extent (self.shape), not the patch-size-extended grid. The
-            # last patch of each axis is padded up to patch_size for the model, but that padded tail lies
-            # OUTSIDE the volume; blending it would over-allocate the accumulator by up to
-            # (patch_size - overlap) per axis (then get cropped away). We crop each patch to its in-volume
-            # part at blend time instead, so nothing outside the volume is ever allocated.
-            n = self._n
-            self._result = torch.zeros(list(layer.shape[:n]) + list(self.shape), dtype=layer.dtype, device=layer.device)
-        self._blend(layer, self.patch_slices[index])
-        self._done[index] = True
-        self._filled += 1
-        return []
-
-    def _blend(self, layer: torch.Tensor, patch_slice: tuple[slice, ...], row_offset: int = 0) -> None:
-        """Blend one patch at its in-volume destination, ``row_offset`` rows down on the first spatial
-        axis (the streaming window origin: 0 for the whole-volume base)."""
-        n = self._n
-        data = layer
-        for dim, s in enumerate(patch_slice):
-            if s.stop - s.start == 1:
-                data = data.unsqueeze(dim=dim + n)
-        # Clamp each spatial destination to the volume and crop the patch to it, BEFORE weighting: the
-        # padded tail of a border patch lies outside the volume, so it has no share of the blend weight
-        # to compute and every index in _weighted_patch stays in range.
-        dest = [slice(s.start, min(s.stop, self.shape[dim])) for dim, s in enumerate(patch_slice)]
-        data = data[tuple([slice(None)] * n + [slice(0, d.stop - d.start) for d in dest])]
-        sweep = self.sweep_axis
-        dest[sweep] = slice(dest[sweep].start - row_offset, dest[sweep].stop - row_offset)
-        slices_dest = tuple([slice(cast(torch.Tensor, self._result).shape[i]) for i in range(n)] + dest)
-        result = cast(torch.Tensor, self._result)
-        lead = slices_dest[:n]
-        if self.patch_combine is None:
-            result[slices_dest] = data
-        elif self.patch_combine.selects:
-            # The kept regions partition the volume: nothing to weight, nothing to sum. Writing the box
-            # this patch owns IS the operation, and it is what carries a discrete output through,
-            # where a weighting would invent values between its classes.
-            box = self._kept_box(patch_slice)
-            kept = [slice(d.start + b.start, d.start + b.stop) for d, b in zip(dest, box, strict=True)]
-            result[(*lead, *kept)] = data[(*lead, *box)]
-        else:
-            result[slices_dest] += self._weighted_patch(data, patch_slice)
-
-    def _kept_box(self, patch_slice: tuple[slice, ...]) -> tuple[slice, ...]:
-        """The sub-box a selection keeps of this patch: the run of ones in its window, per axis.
-
-        Pure geometry, so it is read once from the host-side windows and cached per grid position
-        along each axis. Deriving it from the share instead would read a device tensor: a host sync
-        on every patch, which costs more than the weighting it replaces.
-        """
-        return tuple(self._kept_span(dim, s.start) for dim, s in enumerate(patch_slice))
-
-    def _kept_span(self, dim: int, start: int) -> slice:
-        key = (dim, start)
-        if key not in self._kept:
-            windows, _ = self._weight_geometry()
-            kept = windows[dim][self._position(dim, start)].nonzero().flatten()
-            self._kept[key] = slice(int(kept[0]), int(kept[-1]) + 1)
-        return self._kept[key]
-
-    def _weight_geometry(self) -> tuple[list[list[torch.Tensor]], list[torch.Tensor]]:
-        """Per axis: the blend window, and its sum over the patch grid.
-
-        The outer product of those sums is the total weight covering each voxel. It factorises because
-        the patch grid is a full per-axis product and the window is itself separable
-        (``sum_p prod_d w_d == prod_d sum_k w_d``), so the total is one vector per axis and never exists
-        as a volume, on a 320-row window of a 1331x1775 volume, 13 KB instead of 2.8 GB.
-
-        The patch extent comes from the slices, not from the window: a free axis carries a single
-        broadcast entry (the ModelPatch blend-window contract) that must cover the whole extent.
-
-        The window is asked for per grid position, because a selection opens its border patches to the
-        volume edge (see Trim); a weighting returns the same window everywhere.
-        """
-        if self._geometry is None:
-            combine = cast(PathCombine, self.patch_combine)
-            windows, totals = [], []
-            for dim, extent in enumerate(self.shape):
-                starts = list(self._positions()[dim])
-                length = self.patch_slices[0][dim].stop - self.patch_slices[0][dim].start
-                per_position, total = [], torch.zeros(extent)
-                for position, start in enumerate(starts):
-                    window = combine.window(dim, position, len(starts))
-                    window = window.expand(length) if window.numel() == 1 else window
-                    stop = min(start + length, extent)
-                    total[start:stop] += window[: stop - start]
-                    per_position.append(window)
-                windows.append(per_position)
-                totals.append(total)
-            self._geometry = (windows, totals)
-        return self._geometry
-
-    def _positions(self) -> list[dict[int, int]]:
-        """Per axis, the patch grid's starts in order, each mapped to its position along the axis.
-
-        One pass over the slices for the accumulator's life. Rebuilding the sorted starts per patch
-        made every lookup O(P): 0.07 ms per patch at P = 1331 and 15.6 s over a case of 18,000
-        thin 2.5D patches, against 16 ms for that case here.
-        """
-        if self._grid is None:
-            self._grid = [
-                {
-                    start: position
-                    for position, start in enumerate(sorted({patch[dim].start for patch in self.patch_slices}))
-                }
-                for dim in range(len(self.shape))
-            ]
-        return self._grid
-
-    def _position(self, dim: int, start: int) -> int:
-        return self._positions()[dim][start]
-
-    def _share(self, dim: int, start: int, data: torch.Tensor) -> torch.Tensor | None:
-        """This patch's fraction of the blend weight along one axis, ``w / sum_k w``, cached per axis.
-
-        ``None`` where that fraction is exactly one on every voxel of the axis, which is what an axis
-        holding a single grid position gives (the patch is the whole of the weight there, so the total
-        IS its own window): the blend then skips a full pass over the patch, bit for bit the same
-        values. A patch grid that tiles one axis and spans the other two is that case twice over.
-        """
-        extent = data.shape[self._n + dim]
-        key = (dim, start, extent, data.dtype, data.device)
-        if key not in self._shares:
-            windows, totals = self._weight_geometry()
-            window = windows[dim][self._position(dim, start)]
-            share = window[:extent] / totals[dim][start : start + extent]
-            share = share.to(device=data.device, dtype=data.dtype)
-            self._shares[key] = None if torch.equal(share, torch.ones_like(share)) else share
-        return self._shares[key]
-
-    def _weighted_patch(self, data: torch.Tensor, patch_slice: tuple[slice, ...]) -> torch.Tensor:
-        """``data`` scaled by its SHARE of the blend weight at each voxel, one axis at a time.
-
-        Normalising per patch rather than dividing the assembled volume by an accumulated weight drops
-        both the spatial-sized weight buffer and the final division pass over every channel. The shares
-        sum to one per voxel by construction (the total is the sum over the same grid), so the blend
-        stays exact, and each factor is a ratio of comparable quantities, so it lives in [0, 1] where
-        the raw product underflows fp16 and needed a floor.
-
-        Only the axes whose share is not identically one are applied (see ``_share``), each one pass
-        over the patch: a grid tiled along one axis and spanning the other two hands the patch back
-        untouched. Into a staging buffer the patches share otherwise, one patch-sized allocation per
-        accumulator instead of per blend, and out of place: the caller's tensor is never touched, so
-        the OOM retry (which re-blends the same patch on the CPU) never re-weights it.
-        """
-        shares = []
-        for dim, s in enumerate(patch_slice):
-            share = self._share(dim, s.start, data)
-            if share is not None:
-                view = [1] * data.ndim
-                view[self._n + dim] = -1
-                shares.append(share.view(view))
-        if not shares:
-            return data
-        if (
-            self._weighted is None
-            or self._weighted.shape != data.shape
-            or self._weighted.dtype != data.dtype
-            or self._weighted.device != data.device
-        ):
-            self._weighted = torch.empty_like(data)
-        torch.mul(data, shares[0], out=self._weighted)
-        for share in shares[1:]:
-            self._weighted.mul_(share)
-        return self._weighted
-
-    def _along_sweep(self, span: slice, lead: int) -> tuple[slice, ...]:
-        """Index a result tensor over ``span`` on the sweep axis and whole on every other."""
-        spatial = [span if dim == self.sweep_axis else slice(None) for dim in range(len(self.shape))]
-        return tuple([slice(None)] * lead + spatial)
-
-    def _sweep_shape(self, extent: int) -> list[int]:
-        """The volume's spatial shape with the sweep axis cut down to ``extent``."""
-        return [extent if dim == self.sweep_axis else size for dim, size in enumerate(self.shape)]
-
-    def is_empty(self) -> bool:
-        """True until the first patch has been blended in (no volume-sized buffer allocated yet)."""
-        return self._result is None
-
-    @property
-    def footprint_shape(self) -> list[int]:
-        """Spatial shape held in memory at once: the whole volume for the base accumulator (overridden
-        by StreamingAccumulator, which keeps only a window). Used to size the blend device."""
-        return self.shape
-
-    def is_full(self) -> bool:
-        # O(1): a running counter avoids re-scanning every slot after each added patch
-        # (re-scanning per patch would be O(P^2) per case).
-        return self._filled == self._count
-
-    def assemble(self) -> torch.Tensor:
-        if self._result is None:
-            raise PatchError(
-                "Accumulator.assemble() was called before any patch was added.",
-                f"Expected up to {self._count} patch(es) via add_layer() before assembling.",
-                "Add at least one patch (and check is_full()) before calling assemble().",
-            )
-        result = self._result
-        # Nothing to normalise: each patch was blended in with its share of the weight, so the shares
-        # already sum to one per voxel. No final crop either: patches are cropped at blend time.
-        self._reset()
-        return result
-
-    def _reset(self) -> None:
-        self._result = None
-        self._weighted = None
-        self._filled = 0
-        self._done = [False] * self._count
-
-
-class StreamingAccumulator(Accumulator):
-    """Accumulator holding only the active window along the first spatial axis; it yields each finalized
-    slab as its patches complete, so peak memory is two patch extents (the window and its slide room).
-
-    The patch-grid order (``get_patch_slices_from_shape`` iterates ``itertools.product`` with the first
-    spatial axis outermost) has patch starts along that axis never decreasing, so when a patch starting
-    at ``z`` arrives, every voxel before ``z`` has already received all of its patches and the region up
-    to ``z`` is final. ``add_layer`` returns those finalized slabs (``assemble()``'s values, from the
-    same blend and weight arithmetic applied slab by slab), and ``finalize()`` flushes the tail.
-    """
-
-    def __init__(
-        self,
-        patch_slices: list[tuple[slice, ...]],
-        patch_size: list[int],
-        patch_combine: PathCombine | None = None,
-        batch: bool = True,
-        sweep_axis: int = 0,
-    ) -> None:
-        super().__init__(patch_slices, patch_size, patch_combine, batch, sweep_axis)
-        sweep = self.sweep_axis
-        self._window = min(
-            max(patch[sweep].stop - patch[sweep].start for patch in self.patch_slices), self.shape[sweep]
-        )
-        starts = [patch[sweep].start for patch in self.patch_slices]
-        if (starts and starts[0] != 0) or any(
-            0 > current - previous or current - previous > self._window for previous, current in pairwise(starts)
-        ):
-            raise PatchError(
-                "StreamingAccumulator requires patches starting at row 0 and ordered by non-decreasing start on"
-                " the first spatial axis, advancing by at most one patch extent per step.",
-                "get_patch_slices_from_shape generates this order; a custom patch source must preserve it.",
-            )
-        self._flushed = 0
-
-    def add_layer(self, index: int, layer: torch.Tensor) -> list[tuple[slice, torch.Tensor]]:
-        if self._done[index]:
-            return []
-        n = self._n
-        patch_slice = self.patch_slices[index]
-        # Correctness rests on patches ARRIVING in non-decreasing axis-0-start order, not just on the
-        # slice list being sorted: a patch whose start is already flushed would write at a negative
-        # window offset (dest[0] below), which torch silently reads from the end -> misplaced data, no
-        # error. Fail loud instead. The prediction loop preserves per-case order (shuffle=False), so
-        # this only guards a future misuse (e.g. a shuffling sampler).
-        if patch_slice[self.sweep_axis].start < self._flushed:
-            raise PatchError(
-                f"StreamingAccumulator received patch start {patch_slice[self.sweep_axis].start} after flushing to "
-                f"{self._flushed}: patches must arrive in non-decreasing first-spatial-axis order.",
-                "Add patches in the order get_patch_slices_from_shape generates them (no shuffling).",
-            )
-        slabs = self._advance_to(patch_slice[self.sweep_axis].start)
-        if self._result is None:
-            self._result = torch.zeros(
-                [*layer.shape[:n], *self._sweep_shape(self._window)],
-                dtype=layer.dtype,
-                device=layer.device,
-            )
-        self._blend(layer, patch_slice, row_offset=self._flushed)
-        self._done[index] = True
-        self._filled += 1
-        return slabs
-
-    def finalize(self) -> list[tuple[slice, torch.Tensor]]:
-        """Flush the remaining window and reset for reuse; call once ``is_full()``."""
-        slabs = self._advance_to(self.shape[self.sweep_axis])
-        self._reset()
-        self._flushed = 0
-        return slabs
-
-    @property
-    def footprint_shape(self) -> list[int]:
-        # Only the window is resident, so the blend-device budget is the window's: a huge volume streams
-        # on the GPU within bounded VRAM. Blend and IEEE-correctly-rounded finalize ops (+, *, /, argmax,
-        # cast) are bit-identical CPU/CUDA; only a transcendental-terminated float output (Softmax/Sigmoid)
-        # can differ by ~1 ULP between a window on the GPU and a whole-volume reference on the CPU.
-        return self._sweep_shape(self._window)
-
-    def assemble(self) -> torch.Tensor:
-        raise PatchError(
-            "StreamingAccumulator does not assemble a whole volume.",
-            "Consume the slabs returned by add_layer() and finalize() instead.",
-        )
-
-    def _advance_to(self, z: int) -> list[tuple[slice, torch.Tensor]]:
-        """Finalize the window up to ``z`` (absolute) and shift the window origin there."""
-        z = min(z, self.shape[self.sweep_axis])
-        if self._result is None or z <= self._flushed:
-            return []
-        n = self._n
-        length = z - self._flushed
-        if length > self._window:
-            raise PatchError(
-                f"StreamingAccumulator asked to finalize {length} rows at once with a {self._window}-row window.",
-                "Patch starts may advance by at most one patch extent per step (checked at construction).",
-            )
-        # Cloned: the window slides over these rows right after, so the slab handed out must not be a
-        # view of it. Nothing else to do: the blend weights already sum to one over these voxels.
-        slab = self._result[self._along_sweep(slice(0, length), n)].clone()
-        keep = self._window - length
-        # .clone(): source and destination views overlap when length < window.
-        self._result[self._along_sweep(slice(0, keep), n)] = self._result[
-            self._along_sweep(slice(length, self._window), n)
-        ].clone()
-        self._result[self._along_sweep(slice(keep, self._window), n)] = 0
-        region = slice(self._flushed, z)
-        self._flushed = z
-        return [(region, slab)]
-
-
-class SlabRegionStream:
-    """Slab in → slab out through one region stage, with bounded lookahead.
-
-    The write mirror of the read dispatcher's single-region rule: finalized slabs arrive in order along
-    the first spatial axis (the :class:`StreamingAccumulator`'s order), and each output region is
-    emitted as soon as the input region it pulls has arrived, so only a sliding window of the input is
-    ever resident. The stage itself is two injected callables, both pure region arithmetic + tensor
-    work, so no stage kind has streaming code of its own:
-
-    - ``pull(target_slices) -> source_slices``: the clamped input region an output region is computed
-      from (a transform's ``stream_region_target``/``stream_region_source``, or a halo enlargement).
-    - ``produce(window, target_slices, source_slices) -> tensor``: the output block for
-      ``target_slices``, given exactly the pulled window.
-
-    The schedule is derived from ``pull`` alone: a probe finds which output axis the input slab axis
-    feeds and in which direction (a mirrored or permuted axis streams too, through the sink's
-    random-access region writes), and emission advances along that axis as far as the arrived input
-    allows. Any per-axis monotone map works; nothing here names a stage.
-    """
-
-    def __init__(
-        self,
-        pull: Callable[[tuple[slice, ...]], list[slice]],
-        produce: Callable[[torch.Tensor, tuple[slice, ...], list[slice]], torch.Tensor],
-        in_spatial_shape: list[int],
-        out_spatial_shape: list[int],
-    ) -> None:
-        self._pull = pull
-        self._produce = produce
-        self._in_shape = [int(s) for s in in_spatial_shape]
-        self._out_shape = [int(s) for s in out_spatial_shape]
-        self._axis, self._ascending = self._probe_axis()
-        self._slabs: list[tuple[int, torch.Tensor]] = []
-        self._complete_to = 0
-        self._emitted = 0
-
-    def _probe_axis(self) -> tuple[int, bool]:
-        """Which output axis the input slab axis feeds, and whether in ascending order.
-
-        Probing one output row per axis against the full-region pull identifies the axis whose region
-        controls input axis 0; comparing the first and last rows' pulls gives the direction. A wrong
-        pick can never corrupt the output (emission is gated on the pull of the real regions); it only
-        buffers more, so a map no probe can tell apart falls back to axis 0 ascending.
-        """
-        full = tuple(slice(0, n) for n in self._out_shape)
-        baseline = self._pull(full)[0]
-        for axis, extent in enumerate(self._out_shape):
-            probe = list(full)
-            probe[axis] = slice(0, 1)
-            first = self._pull(tuple(probe))[0]
-            if (first.start, first.stop) == (baseline.start, baseline.stop):
-                continue
-            probe[axis] = slice(extent - 1, extent)
-            last = self._pull(tuple(probe))[0]
-            return axis, first.start <= last.start
-        return 0, True
-
-    def _prefix_slices(self, m: int) -> tuple[slice, ...]:
-        """The first ``m`` output rows in iteration order, as absolute region slices."""
-        extent = self._out_shape[self._axis]
-        rows = slice(0, m) if self._ascending else slice(extent - m, extent)
-        slices = [slice(0, n) for n in self._out_shape]
-        slices[self._axis] = rows
-        return tuple(slices)
-
-    def push(self, region: slice, slab: torch.Tensor) -> list[tuple[tuple[slice, ...], torch.Tensor]]:
-        """Take the next finalized slab (rows ``region`` of the input) and emit what it completes."""
-        if region.start != self._complete_to:
-            raise PatchError(
-                f"SlabRegionStream received rows [{region.start}, {region.stop}) after "
-                f"[0, {self._complete_to}): slabs must arrive contiguously from the first row.",
-                "Push the slabs exactly as the StreamingAccumulator yields them.",
-            )
-        self._slabs.append((region.start, slab))
-        self._complete_to = region.stop
-        return self._emit()
-
-    def finalize(self) -> list[tuple[tuple[slice, ...], torch.Tensor]]:
-        """Emit whatever the completed input still allows and verify nothing is left behind."""
-        emitted = self._emit()
-        if self._emitted != self._out_shape[self._axis]:
-            raise PatchError(
-                f"SlabRegionStream finalized with {self._emitted} of {self._out_shape[self._axis]} "
-                f"output rows emitted (input complete to {self._complete_to} of {self._in_shape[0]}).",
-                "Push every slab of the input before finalizing.",
-            )
-        self._slabs.clear()
-        return emitted
-
-    def _emit(self) -> list[tuple[tuple[slice, ...], torch.Tensor]]:
-        extent = self._out_shape[self._axis]
-        # The pull of an iteration prefix grows monotonically with it, so the furthest emittable row is
-        # a binary search. O(log rows) pull calls per push, and a pull may be more than slice
-        # arithmetic (a declaration may copy the attribute it reads).
-        low, high = self._emitted, extent
-        while low < high:
-            middle = (low + high + 1) // 2
-            if self._pull(self._prefix_slices(middle))[0].stop <= self._complete_to:
-                low = middle
-            else:
-                high = middle - 1
-        m = low
-        if m == self._emitted:
-            return []
-        rows = slice(self._emitted, m) if self._ascending else slice(extent - m, extent - self._emitted)
-        target = list(self._prefix_slices(m))
-        target[self._axis] = rows
-        source = self._pull(tuple(target))
-        window = self._window(source)
-        result = self._produce(window, tuple(target), source)
-        self._emitted = m
-        if m < extent:
-            remaining = [slice(0, n) for n in self._out_shape]
-            remaining[self._axis] = slice(m, extent) if self._ascending else slice(0, extent - m)
-            keep_from = self._pull(tuple(remaining))[0].start
-            self._slabs = [
-                (start, slab) for start, slab in self._slabs if start + slab.shape[-len(self._in_shape)] > keep_from
-            ]
-        else:
-            self._slabs.clear()
-        return [(tuple(target), result)]
-
-    def _window(self, source: list[slice]) -> torch.Tensor:
-        """The buffered input restricted to ``source``: rows gathered from the arrived slabs, the
-        other axes sliced in place."""
-        axis0 = source[0]
-        n_lead = self._slabs[0][1].dim() - len(self._in_shape)
-        lead = (slice(None),) * n_lead
-        pieces = []
-        for start, slab in self._slabs:
-            stop = start + slab.shape[n_lead]
-            lo, hi = max(start, axis0.start), min(stop, axis0.stop)
-            if lo < hi:
-                pieces.append(slab[(*lead, slice(lo - start, hi - start))])
-        window = pieces[0] if len(pieces) == 1 else torch.cat(pieces, dim=n_lead)
-        if window.shape[n_lead] != axis0.stop - axis0.start:
-            raise PatchError(
-                f"SlabRegionStream window rows [{axis0.start}, {axis0.stop}) are not fully buffered.",
-                "The pull map must never reach below rows already pruned or above rows arrived.",
-            )
-        return window[(*lead, slice(None), *source[1:])]
-
-
-class SlabAligner:
-    """Merge several slab streams over the same axis into jointly finalized intervals.
-
-    The cross-stream mirror of :class:`StreamingAccumulator`: each stream (a TTA copy's accumulator)
-    emits finalized slabs in non-decreasing order along the shared first spatial axis, and a consumer
-    that needs every stream's rows together (a cross-copy reduction) can only advance to the
-    slowest frontier. ``push`` takes one stream's new slabs and returns the intervals that just
-    became complete, each carrying every stream's rows; only the inter-stream skew is ever buffered,
-    and a single stream passes through untouched. Nothing here knows what a stream is: it is pure
-    interval arithmetic over ``nb_streams`` ordered emitters.
-    """
-
-    def __init__(self, nb_streams: int, lead_dims: int = 1) -> None:
-        self._nb_streams = nb_streams
-        self._lead = lead_dims
-        self._pending: dict[int, list[tuple[int, torch.Tensor]]] = {}
-        self._frontiers: dict[int, int] = {}
-        self._consumed = 0
-        self._completed: set[int] = set()
-
-    @property
-    def complete(self) -> bool:
-        """Whether every stream has pushed its last slab."""
-        return len(self._completed) == self._nb_streams
-
-    def push(
-        self, stream: int, slabs: list[tuple[slice, torch.Tensor]], finished: bool = False
-    ) -> list[tuple[slice, dict[int, torch.Tensor]]]:
-        """Take ``stream``'s freshly finalized slabs and return the newly joint-final intervals."""
-        pending = self._pending.setdefault(stream, [])
-        for region, slab in slabs:
-            pending.append((region.start, slab))
-            self._frontiers[stream] = region.stop
-        if finished:
-            self._completed.add(stream)
-        if len(self._frontiers) < self._nb_streams:
-            return []
-        joint = min(self._frontiers.values())
-        if joint <= self._consumed:
-            return []
-        lead = (slice(None),) * self._lead
-        rows: dict[int, torch.Tensor] = {}
-        for key in sorted(self._pending):
-            pieces = [
-                slab[
-                    (
-                        *lead,
-                        slice(max(start, self._consumed) - start, min(start + slab.shape[self._lead], joint) - start),
-                    )
-                ]
-                for start, slab in self._pending[key]
-                if start < joint and start + slab.shape[self._lead] > self._consumed
-            ]
-            if not pieces:
-                raise PatchError(
-                    f"SlabAligner stream {key} holds no rows for [{self._consumed}, {joint}).",
-                    "Push the slabs exactly as the accumulators finalize them, without skipping a stream.",
-                )
-            rows[key] = pieces[0] if len(pieces) == 1 else torch.cat(pieces, dim=self._lead)
-            self._pending[key] = [
-                (start, slab) for start, slab in self._pending[key] if start + slab.shape[self._lead] > joint
-            ]
-        interval = slice(self._consumed, joint)
-        self._consumed = joint
-        return [(interval, rows)]
-
-
-class _PatchGrid(NamedTuple):
-    """One copy's cut: the axis the patches are ordered by, and the patches themselves."""
-
-    sweep_axis: int
-    slices: list[tuple[slice, ...]]
-
-
-class Patch(ABC):
-    """Abstract base class for dataset-level and model-level patch definitions."""
-
-    @abstractmethod
-    def __init__(
-        self,
-        patch_size: list[int],
-        overlap: OverlapSpec = None,
-        pad_value: float | None = 0,
-        extend_slice: int = 0,
-    ) -> None:
-        if extend_slice != 0 and patch_size is not None and patch_size[0] != 1:
-            raise ValueError(
-                "`extend_slice` can only be used when patch_size[0] == 1 "
-                f"(got patch_size[0]={patch_size[0]}, extend_slice={extend_slice})"
-            )
-        self.patch_size = patch_size
-        self.overlap = overlap
-        if isinstance(self.overlap, int):
-            if self.overlap < 0:
-                self.overlap = None
-        #: The spatial shape each copy's grid is cut on, recorded by ``load``.
-        self._shapes: dict[int, list[int]] = {}
-        #: The cut itself, made on first use and left out of the pickle (``__getstate__``).
-        self._grids: dict[int, _PatchGrid] = {}
-        self.pad_value = pad_value
-        self.extend_slice = extend_slice
-        # Models need every patch at the declared size, so the last patch of an axis is padded up to it.
-        # A consumer that REDUCES patches instead (streamed evaluation) must see only in-volume voxels:
-        # padded ones would pollute its running sums, so it turns this off and takes the cropped patch.
-        self.pad_to_patch = True
-        #: Voxels of context read past each face of a grid patch, clamped to the volume, for a consumer
-        #: that reduces patches but scores through a window (a metric's halo). The grid keeps its
-        #: disjoint slots (``get_patch_slices``); ``core_in_read`` says where a slot sits in its read.
-        #: Unpadded only (``pad_to_patch`` False): a model input padded to the patch has no core.
-        self.halo = 0
-        # The model's per-axis downsampling factor a FREE (``0``) axis rounds up to, set before the grids
-        # are cut so each case's whole-axis extent lands on a valid model input. ``None`` outside a model
-        # (evaluation) or for a network that never downsamples.
-        self.free_axis_multiple: list[int] | None = None
-        # Whether a free (``0``) axis was DECLARED. Captured now because the OOM re-plan later pins
-        # ``patch_size`` to a concrete size in place, erasing the ``0`` the overlap default keys on.
-        self._declared_free_axis: bool = (
-            patch_size is not None and any(p == 0 for p in patch_size) and not all(p == 0 for p in patch_size)
-        )
-
-    def load(self, shape: list[int], a: int = 0) -> None:
-        """Record the spatial shape copy ``a``'s grid is cut on; the cut itself waits for a consumer."""
-        self._shapes[a] = list(shape)
-        self._grids.pop(a, None)
-
-    def _grid(self, a: int) -> _PatchGrid:
-        """Copy ``a``'s cut, made once. A pure function of the recorded shape and this patch's own
-        configuration, so a rank rebuilds it instead of unpickling it (``__getstate__``)."""
-        grid = self._grids.get(a)
-        if grid is None:
-            shape = self._shapes[a]
-            # The grid decides its own sweep axis and the reassembly reads it back (get_sweep_axis): one
-            # source of truth, because a grid emitted for one axis and reassembled along another hands out
-            # regions that are not final, with nothing to report it.
-            sweep_axis = best_sweep_axis(concretize_patch_size(self.patch_size, shape, self.free_axis_multiple), shape)
-            slices = get_patch_slices_from_shape(
-                self.patch_size,
-                shape,
-                self.overlap,
-                self.free_axis_multiple,
-                self._declared_free_axis,
-                sweep_axis,
-            )
-            grid = self._grids[a] = _PatchGrid(sweep_axis, slices)
-        return grid
-
-    def __getstate__(self) -> dict[str, Any]:
-        """The cuts stay out of the pickle: ``mp.spawn`` pickles the configured object once per
-        rank, and every manager's per-case grids dominated its bytes. Only the shapes travel, and
-        each rank cuts the same grids back from them on first use.
-
-        Cutting costs more than unpickling would have (100 cases of 2048 patches: 31 ms against 81
-        ms in process); it is the transfer, paid once per rank, that dominates. On a 2-rank CPU
-        spawn of 1000 such cases the payload goes from 35.2 to 0.44 MiB and the second rank holds
-        every grid 3.0 s after the launcher's stamp instead of 4.8 s.
-        """
-        return {**self.__dict__, "_grids": {}}
-
-    def get_sweep_axis(self, a: int = 0) -> int:
-        """The axis this grid is ordered by, and so the one reassembly must slide along."""
-        return self._grid(a).sweep_axis
-
-    @abstractmethod
-    def init(self, key: str):
-        pass
-
-    def get_patch_slices(self, a: int = 0) -> list[tuple[slice, ...]]:
-        return self._grid(a).slices
-
-    def read_slices(self, a: int, index: int, shape: Sequence[int]) -> list[slice]:
-        """The region patch ``index`` of copy ``a`` reads: its grid slot, widened by the halo and
-        clamped to the volume (``shape`` may carry leading non-spatial axes)."""
-        slot = self._grid(a).slices[index]
-        if not self.halo:
-            return list(slot)
-        spatial = [int(extent) for extent in shape[len(shape) - len(slot) :]]
-        return _HaloPull([self.halo] * len(slot), spatial)(slot)
-
-    def core_in_read(self, a: int, index: int) -> tuple[slice, ...]:
-        """Where patch ``index``'s grid slot sits within its read: the halo in from each face, less
-        where the volume's own face cut the halo short."""
-        return tuple(
-            slice(min(self.halo, s.start), min(self.halo, s.start) + s.stop - s.start)
-            for s in self._grid(a).slices[index]
-        )
-
-    def get_read_plan(
-        self, data_shape: list[int] | tuple[int, ...], index: int, a: int, is_input: bool
-    ) -> PatchReadPlan:
-        slot = self.read_slices(a, index, data_shape)
-        slices_pre = [slice(None) for _ in data_shape[: -len(slot)]]
-        extend_slice = self.extend_slice if is_input else 0
-
-        bottom = extend_slice // 2
-        top = int(np.ceil(extend_slice / 2))
-        s = slice(
-            (slot[0].start - bottom if slot[0].start - bottom >= 0 else 0),
-            (slot[0].stop + top if slot[0].stop + top <= data_shape[len(slices_pre)] else data_shape[len(slices_pre)]),
-        )
-        slices = [s, *slot[1:]]
-        reflect_padding = [0 for _ in range((len(slices) - 1) * 2)] + [0, 0]
-        if extend_slice > 0 and (s.stop - s.start) < bottom + top + 1:
-            if slot[0].start - bottom < 0:
-                reflect_padding[-2] = bottom - slot[0].start
-            if slot[0].stop + top > data_shape[len(slices_pre)]:
-                reflect_padding[-1] = slot[0].stop + top - data_shape[len(slices_pre)]
-
-        constant_padding = []
-        if self.pad_to_patch and self.patch_size is not None:
-            nspatial = len(slices)
-            for dim_it, _slice in enumerate(reversed(slices)):
-                axis = nspatial - 1 - dim_it
-                extent = data_shape[-dim_it - 1]
-                declared = self.patch_size[axis]
-                if declared != 0:
-                    target = declared
-                else:
-                    # A FREE axis pads up to THIS case's extent rounded to the model's downsampling
-                    # multiple, so a small heterogeneous case still reaches the network at a valid input
-                    # size (the up-front worst-case sizing only guarantees the largest case).
-                    m = free_axis_rounding(self.free_axis_multiple, axis, nspatial)
-                    target = ((extent + m - 1) // m) * m if m > 1 else extent
-                p = 0 if _slice.start + target <= extent else target - (extent - _slice.start)
-                constant_padding.append(0)
-                constant_padding.append(p)
-
-        return PatchReadPlan(
-            data_slices=tuple(slices_pre + slices),
-            reflect_padding=tuple(reflect_padding),
-            constant_padding=tuple(constant_padding),
-            concatenate_extend_slice=extend_slice > 0,
-        )
-
-    def apply_read_plan(self, data: torch.Tensor, plan: PatchReadPlan) -> torch.Tensor:
-        data_sliced = data
-        if any(plan.reflect_padding):
-            data_sliced = F.pad(data_sliced, plan.reflect_padding, "reflect")
-        if any(plan.constant_padding):
-            data_sliced = self._pad_constant(data_sliced, plan.constant_padding)
-        if self.patch_size is not None and not all(p == 0 for p in self.patch_size):
-            for d in [i for i, v in enumerate(reversed(self.patch_size)) if v == 1]:
-                data_sliced = torch.squeeze(data_sliced, dim=len(data_sliced.shape) - d - 1)
-        return (
-            torch.cat([data_sliced[:, i, ...] for i in range(data_sliced.shape[1])], dim=0)
-            if plan.concatenate_extend_slice
-            else data_sliced
-        )
-
-    def _pad_constant(self, data: torch.Tensor, padding: tuple[int, ...]) -> torch.Tensor:
-        """``data`` padded (``F.pad`` pair order) with the configured value, a uint8 map with zero,
-        and anything else with its own minimum when no value is configured.
-
-        That minimum never leaves the device: the bands are filled from the 0-d tensor after a zero
-        pad, so a padded patch costs no host round trip (37 of the 64 patches of a 100^3 case at
-        32^3 are padded, measured).
-        """
-        if data.dtype == torch.uint8:
-            return F.pad(data, padding, "constant", 0)
-        if self.pad_value is not None:
-            return F.pad(data, padding, "constant", self.pad_value)
-        padded = F.pad(data, padding, "constant", 0)
-        lowest = data.min()
-        for pair, (before, after) in enumerate(zip(padding[::2], padding[1::2], strict=True)):
-            dim = padded.dim() - 1 - pair
-            if before:
-                padded.narrow(dim, 0, before).copy_(lowest)
-            if after:
-                padded.narrow(dim, padded.shape[dim] - after, after).copy_(lowest)
-        return padded
-
-    def get_data(self, data: torch.Tensor, index: int, a: int, is_input: bool) -> list[torch.Tensor]:
-        plan = self.get_read_plan(list(data.shape), index, a, is_input)
-        data_sliced = data[plan.data_slices]
-        return self.apply_read_plan(data_sliced, plan)
-
-    def get_size(self, a: int = 0) -> int:
-        return len(self._grid(a).slices)
-
-
-@config("Patch")
-class DatasetPatch(Patch):
-    """Patch definition applied when sampling data from datasets."""
-
-    def __init__(
-        self,
-        patch_size: list[int] = [128, 128, 128],
-        overlap: OverlapSpec = None,
-        pad_value: float | None = None,
-        extend_slice: int = 0,
-    ) -> None:
-        super().__init__(patch_size, overlap, pad_value, extend_slice)
-
-    def init(self, key: str = ""):
-        pass
-
-
-@config()
-class ModelPatch(Patch):
-    """Patch definition applied inside model graphs during prediction or training."""
-
-    def __init__(
-        self,
-        patch_size: list[int] = [128, 128, 128],
-        overlap: OverlapSpec = None,
-        patch_combine: str | None = None,
-        pad_value: float | None = None,
-        extend_slice: int = 0,
-    ) -> None:
-        super().__init__(patch_size, overlap, pad_value, extend_slice)
-        self._patch_combine = patch_combine
-        self.patch_combine: PathCombine | None = None
-
-    def init(self, key: str):
-        if self._patch_combine is not None:
-            module, name = get_module(self._patch_combine, "konfai.data.patching")
-            self.patch_combine = apply_config(key)(getattr(module, name))()
-        if self.patch_size is not None and self.overlap is not None:
-            if self.patch_combine is not None:
-                kept = blend_axes(self.patch_size)
-                self.patch_combine.set_patch_config(kept, blend_overlap(self.overlap, kept))
-        else:
-            self.patch_combine = None
-
-    def disassemble(self, *data_list: torch.Tensor) -> Iterator[list[torch.Tensor]]:
-        for i in range(self.get_size()):
-            yield [self.get_data(data, i, 0, True) for data in data_list]
+from konfai.utils.errors import DatasetManagerError, PatchError
+from konfai.utils.utils import env_flag
 
 
 class DatasetManager:
@@ -2083,7 +201,7 @@ class DatasetManager:
         # from the source and each stream's finalize renames over the old entry. Never the default --
         # the boundary IS the per-case resume.
         self._rewrite_saves = False
-        # The per-rank budget the sweeps size their slabs against (None = the fixed SWEEP_SLAB_ROWS).
+        # The per-rank budget the sweeps size their slabs against (None = the fixed budget.SWEEP_SLAB_ROWS).
         self._sweep_budget_bytes: float | None = None
         # The store's own read granularity, resolved on first use (None is an answer, not a miss).
         self._read_granularity: object = _UNRESOLVED
@@ -3136,8 +1254,12 @@ class DatasetManager:
         blocks are free, and the cap is what bounds them: on a 513x1331x1776 sweep in 40 blocks, a
         second block in flight recovers 0.5 s of a 6.7 s run and a third recovers none.
         """
-        depth = _sweep_pipeline_depth()
-        while depth and depth < _SWEEP_MAX_DEPTH and self._keeps_the_block(spatial, channels, plans, tile, depth + 1):
+        depth = sweep_module._sweep_pipeline_depth()
+        while (
+            depth
+            and depth < budget._SWEEP_MAX_DEPTH
+            and self._keeps_the_block(spatial, channels, plans, tile, depth + 1)
+        ):
             depth += 1
         return depth
 
@@ -3166,7 +1288,7 @@ class DatasetManager:
         plans: Sequence["_ReadStagePlan"] = (),
         depth: int | None = None,
     ) -> int:
-        """The tallest region the sweep will cut whatever the budget: ``SWEEP_SLAB_ROWS`` on a CPU,
+        """The tallest region the sweep will cut whatever the budget: ``budget.SWEEP_SLAB_ROWS`` on a CPU,
         taller on a GPU as its free memory allows. What the budget then affords is
         :meth:`_sweep_tile`'s.
 
@@ -3175,7 +1297,7 @@ class DatasetManager:
         region counted as one landed plane raised the cap by the pull ratio of a GPU ``Resample``,
         and nothing else bounded it where no host budget was declared.
         """
-        cap = max(1, int(SWEEP_SLAB_ROWS))
+        cap = max(1, int(budget.SWEEP_SLAB_ROWS))
         # Never below the store's own block: a region shorter than one reads it whole regardless
         # (the hull is what a chunked read decodes), so cutting under it buys no memory back and
         # only reads the same bytes again for the next region.
@@ -3308,7 +1430,7 @@ class DatasetManager:
         """The tallest region up to ``cap`` rows whose priced block holds inside ``budget``, ``1``
         when none does: the one search both ceilings (the rank's budget, the device's free memory)
         are answered by."""
-        depth = _sweep_pipeline_depth() if depth is None else depth
+        depth = sweep_module._sweep_pipeline_depth() if depth is None else depth
         low, high = 1, max(1, int(cap))
         while low < high:
             middle = (low + high + 1) // 2
@@ -3448,7 +1570,7 @@ class DatasetManager:
         landed one, or a stage declaring eight volumes-worth of buffers, costs what it costs. The
         search is over the height, because that is the one free parameter of the decomposition.
         """
-        depth = _sweep_pipeline_depth() if depth is None else depth
+        depth = sweep_module._sweep_pipeline_depth() if depth is None else depth
         cap = self._sweep_rows(spatial, channels, plans, depth)
         budget = self._sweep_budget_bytes
         if not budget or budget <= 0:
