+++ tests/unit/test_case_reduction.py
@@ -26,7 +26,6 @@ import weakref
 from pathlib import Path
 from types import SimpleNamespace
 
-import konfai.data.patching as patching_module
 import konfai.data.transform as transform_module
 import numpy as np
 import pytest
@@ -880,13 +879,13 @@ def test_a_host_chain_gets_a_meter_and_a_kernel_without_one_gets_none(monkeypatc
     """A host chain has no allocator counter, and the kernel's resident peak is the meter it does
     have. A kernel that offers no reset gets no meter at all, rather than a run-long peak read as
     if it were one region's."""
-    monkeypatch.setattr(patching_module, "reset_resident_peak", lambda: True)
-    monkeypatch.setattr(patching_module, "resident_bytes", lambda: 1_000_000)
-    monkeypatch.setattr(patching_module, "peak_resident_bytes", lambda: 1_500_000)
+    monkeypatch.setattr("konfai.data.patching.budget.reset_resident_peak", lambda: True)
+    monkeypatch.setattr("konfai.data.patching.budget.resident_bytes", lambda: 1_000_000)
+    monkeypatch.setattr("konfai.data.patching.budget.peak_resident_bytes", lambda: 1_500_000)
     meter = open_held_meter(None)
     assert meter is not None and meter.held() == 500_000
 
-    monkeypatch.setattr(patching_module, "reset_resident_peak", lambda: False)
+    monkeypatch.setattr("konfai.data.patching.budget.reset_resident_peak", lambda: False)
     assert open_held_meter(None) is None
