+++ tests/unit/test_save_streaming.py
@@ -109,9 +109,8 @@ def test_multi_slab_sweep_writes_the_same_cache_as_the_whole_volume_load(
 ) -> None:
     """Three slabs instead of one: slab targets at non-zero offsets, the header written on the first
     slab only, and a region stage composed across slab boundaries."""
-    from konfai.data import patching as patching_module
 
-    monkeypatch.setattr(patching_module, "SWEEP_SLAB_ROWS", 4)
+    monkeypatch.setattr("konfai.data.patching.budget.SWEEP_SLAB_ROWS", 4)
     source = _source(tmp_path)
     classic = [Permute("2|1|0"), Clip(0.0, 50.0), Save(str(tmp_path / "cache_classic"))]
     swept = [Permute("2|1|0"), Clip(0.0, 50.0), Save(str(tmp_path / "cache_swept"))]
@@ -129,10 +128,9 @@ def test_multi_slab_sweep_writes_the_same_cache_as_the_whole_volume_load(
 
 def test_failed_multi_slab_sweep_leaves_no_partial_cache(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """A failure after the first slab is written must remove the partial entry, not publish it."""
-    from konfai.data import patching as patching_module
     from konfai.utils import dataset as dataset_module
 
-    monkeypatch.setattr(patching_module, "SWEEP_SLAB_ROWS", 4)
+    monkeypatch.setattr("konfai.data.patching.budget.SWEEP_SLAB_ROWS", 4)
     source = _source(tmp_path)
     manager = _manager(source, [Clip(0.0, 50.0), Save(str(tmp_path / "cache"))])
     reference = _whole_volume_patches(manager, [Clip(0.0, 50.0)])
